+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/ViewModels/IIsDirty.cs
@@ -1,7 +1,0 @@
-﻿namespace TD.WPF.ViewModels
-{
-    public interface IIsDirty
-    {
-        bool IsDirty { get; }
-    }
-}