+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/Views/Maintenance/SymbolCreate.xaml.cs
@@ -1,0 +1,57 @@
+﻿using System;
+using System.Windows;
+using TD.Models;
+using TD.WPF.ViewModels.Maintenance;
+
+namespace TD.WPF.Views.Maintenance
+{
+    /// <summary>
+    /// Interaction logic for CreateSymbol.xaml
+    /// </summary>
+    public partial class SymbolCreate : Window
+    {
+        private readonly SymbolCreateVM _viewModel;
+
+        public Symbol Symbol => _viewModel.Symbol;
+
+        public SymbolCreate()
+        {
+            InitializeComponent();
+
+            _viewModel = new SymbolCreateVM();
+            DataContext = _viewModel;
+            SourceInitialized += CreateSymbol_SourceInitialized;
+        }
+
+        private async void CreateSymbol_SourceInitialized(object? sender, EventArgs e)
+        {
+            try
+            {
+                await _viewModel.LoadExchangesAsync();
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show($"Load exchanges failed.\n{ex.Message}", "Ozz Trade Diary", MessageBoxButton.OK, MessageBoxImage.Error);
+            }
+
+            try
+            {
+                await _viewModel.LoadCurrenciesAsync();
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show($"Load currencies failed.\n{ex.Message}", "Ozz Trade Diary", MessageBoxButton.OK, MessageBoxImage.Error);
+            }
+        }
+
+        private void OkButton_Click(object sender, RoutedEventArgs e)
+        {
+            DialogResult = true;
+        }
+
+        private void MarketTypeComboBox_DropDownClosed(object sender, EventArgs e)
+        {
+            TickerTextBox.Focus();
+        }
+    }
+}