+++ OzzTradeDiary.NET/OzzTradeDiary/Extensions/EnumExtension.cs
@@ -1,0 +1,59 @@
+﻿using System.ComponentModel.DataAnnotations;
+using System.Reflection;
+using System.Resources;
+
+namespace TD.Extensions
+{
+    public static class EnumExtension
+    {
+        /// <summary>
+        ///     A generic extension method that aids in reflecting 
+        ///     and retrieving any attribute that is applied to an `Enum`.
+        /// </summary>
+        public static TAttribute? GetAttribute<TAttribute>(this Enum enumValue)
+                where TAttribute : Attribute
+        {
+            var attribute = enumValue.GetType()
+                            .GetMember(enumValue.ToString())
+                            .First()
+                            .GetCustomAttribute<TAttribute>();
+
+            return attribute;
+        }
+
+        public static string GetDisplayValue(this Enum value)
+        {
+            var fieldInfo = value.GetType().GetField(value.ToString());
+
+            DisplayAttribute[] descriptionAttributes = Array.Empty<DisplayAttribute>();
+            if (fieldInfo != null)
+                descriptionAttributes = fieldInfo.GetCustomAttributes(typeof(DisplayAttribute), false) as DisplayAttribute[]
+                                     ?? Array.Empty<DisplayAttribute>();
+
+            if (descriptionAttributes.Length > 0 && descriptionAttributes[0].ResourceType != null)
+                return LookupResource(descriptionAttributes[0].ResourceType, descriptionAttributes[0].Name);
+
+            if (descriptionAttributes == null) return string.Empty;
+            return (descriptionAttributes.Length > 0) ? descriptionAttributes[0].Name : value.ToString();
+        }
+
+        public static IEnumerable<T> GetValues<T>()
+        {
+            return Enum.GetValues(typeof(T)).Cast<T>();
+        }
+
+        private static string LookupResource(Type resourceManagerProvider, string resourceKey)
+        {
+            foreach (PropertyInfo staticProperty in resourceManagerProvider.GetProperties(BindingFlags.Static | BindingFlags.NonPublic | BindingFlags.Public))
+            {
+                if (staticProperty.PropertyType == typeof(ResourceManager))
+                {
+                    var resourceManager = (ResourceManager)staticProperty.GetValue(null, null);
+                    return resourceManager.GetString(resourceKey);
+                }
+            }
+
+            return resourceKey; // Fallback with the key name
+        }
+    }
+}